+++ eventsourcing/examples/aggregate6/domainmodel.py
@@ -1,11 +1,9 @@
 from dataclasses import dataclass
 from datetime import datetime, timezone
 from functools import singledispatch
-from time import monotonic
-from typing import Iterable, Optional, Tuple, TypeVar, Union
+from typing import Callable, Iterable, Optional, Tuple, TypeVar, Union
 from uuid import UUID, uuid4
 
-from eventsourcing.application import MutatorFunction, ProjectorFunction
 from eventsourcing.domain import Snapshot
 
 
@@ -17,7 +15,7 @@ class DomainEvent:
 
 
 def create_timestamp() -> datetime:
-    return datetime.fromtimestamp(monotonic(), timezone.utc)
+    return datetime.now(tz=timezone.utc)
 
 
 @dataclass(frozen=True)
@@ -28,11 +26,12 @@ class Aggregate:
 
 
 TAggregate = TypeVar("TAggregate", bound=Aggregate)
+MutatorFunction = Callable[..., Optional[TAggregate]]
 
 
 def aggregate_projector(
     mutator: MutatorFunction[TAggregate],
-) -> ProjectorFunction[TAggregate]:
+) -> Callable[[Optional[TAggregate], Iterable[DomainEvent]], Optional[TAggregate]]:
     def project_aggregate(
         aggregate: Optional[TAggregate], events: Iterable[DomainEvent]
     ) -> Optional[TAggregate]: